fix(base.controller): set resource type (#21)
* feat(formatResponse): specify model type in JSON payload

The formatResponse utility tool now adds the model type to the JSON
payload. This is required by the JSON API spec.

* fix(mot): specify type in JSON payload

The resource type, 'mots', will now be added to each JSOn object
returned from API calls. This is required by the JSON API spec used by
the serializers in the Ember frontend.